+++ services/client/src/components/About.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const About = () => (
+    <div>
+        <h1 className="title is-1">About</h1>
+        <hr/><br/>
+        <p>Add something relevant here.</p>
+    </div>
+)
+
+export default About;
